Update 7/20/2020
Updated:
The application is more prepared to handle processing a Contract object and outputting it as a ContractStage object.

There are runtime errors that need to be addressed still in the job-context.xml file. Possible resolution: create the steps and item readers for the beans that are giving me trouble.